hw 2 grading

+++ src/homework/02_decisions/decisions.h
@@ -8,5 +8,5 @@ int get_grade_points(std::string letter_grade);
 
 //Write a function prototype with name calculate_gpa that accepts an int named credit_hours and
 //an double named credit_points returns double.  
-double calculate_gpa(int credit_hours, double credit_points);
+double calculate_gpa(double credit_hours, double credit_points);
